requirements.txt

streamlit>=1.35.0
PyPDF2>=3.0.1
docx2txt>=0.8
pandas>=2.2.2
numpy>=1.26.4
rapidfuzz>=3.9.7
scikit-learn>=1.4.2
sentence-transformers>=3.0.1
python-dotenv>=1.0.1
matplotlib>=3.8.4
google-generativeai>=0.5.4